# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AA PvE API",
        "version": "0.0.1",
        "description": ""
    },
    "paths": {
...
        "/pve/api/rotations/{rotation_id}/entries/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_rotation_entries",
                "summary": "Get Rotation Entries",
                "parameters": [
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EntrySchema"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "CanAccessPVE": []
                    }
                ]
            },
            "post": {
                "operationId": "allianceauth_pve_api_entries_new_entry",
                "summary": "New Entry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntryFormErrorsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EntryFormSchema"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "NeedsPermission": []
                    }
                ]
            }
        },
        "/pve/api/rotations/{rotation_id}/entries/{entry_id}/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_rotation_entry",
                "summary": "Get Rotation Entry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntryDetailsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "CanAccessPVE": []
                    }
                ]
            },
            "delete": {
                "operationId": "allianceauth_pve_api_entries_delete_rotation_entry",
                "summary": "Delete Rotation Entry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "NeedsPermission": []
                    }
                ]
            },
            "post": {
                "operationId": "allianceauth_pve_api_entries_edit_entry",
                "summary": "Edit Entry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntryFormErrorsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EntryFormSchema"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "NeedsPermission": []
                    }
                ]
            }
        },
        "/pve/api/rotations/{rotation_id}/entries/{entry_id}/roles/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_rotation_entry_roles",
                "summary": "Get Rotation Entry Roles",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EntryRoleSchema"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "CanAccessPVE": []
                    }
                ]
            }
        },
        "/pve/api/rotations/{rotation_id}/entries/{entry_id}/shares/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_rotation_entry_shares",
                "summary": "Get Rotation Entry Shares",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EntryCharacterSchema"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "CanAccessPVE": []
                    }
                ]
            }
        },
        "/pve/api/rotations/{rotation_id}/entries/{entry_id}/items/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_rotation_entry_items",
                "summary": "Get Rotation Entry Items",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ExtendedEntryItemSchema"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "CanAccessPVE": []
                    }
                ]
            }
        },
        "/pve/api/rotations/{rotation_id}/entries/{entry_id}/edit/": {
            "get": {
                "operationId": "allianceauth_pve_api_entries_get_entry_for_edit",
                "summary": "Get Entry For Edit",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "rotation_id",
                        "schema": {
                            "title": "Rotation Id",
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExtendedEntryFormSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "rotations",
                    "entries"
                ],
                "security": [
                    {
                        "NeedsPermission": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EntrySchema": {
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "created_at": {
                        "format": "date-time",
                        "title": "Created At",
                        "type": "string"
                    },
                    "created_by_character": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EveCharacterSchema"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "total_user_count": {
                        "title": "Total User Count",
                        "type": "integer"
                    },
                    "estimated_total": {
                        "title": "Estimated Total",
                        "type": "integer"
                    },
                    "estimated_total_after_tax": {
                        "title": "Estimated Total After Tax",
                        "type": "number"
                    },
                    "actual_total_after_tax": {
                        "title": "Actual Total After Tax",
                        "type": "number"
                    },
                    "actual_total_from_items": {
                        "title": "Actual Total From Items",
                        "type": "number"
                    }
                },
                "required": [
                    "id",
                    "created_at",
                    "created_by_character",
                    "total_user_count",
                    "estimated_total",
                    "estimated_total_after_tax",
                    "actual_total_after_tax",
                    "actual_total_from_items"
                ],
                "title": "EntrySchema",
                "type": "object"
            },
...
            "EntryFormErrorsSchema": {
                "properties": {
                    "estimated_total": {
                        "default": [],
                        "items": {
                            "type": "string"
                        },
                        "title": "Estimated Total",
                        "type": "array"
                    },
                    "funding_project_id": {
                        "default": [],
                        "items": {
                            "type": "string"
                        },
                        "title": "Funding Project Id",
                        "type": "array"
                    },
                    "funding_percentage": {
                        "default": [],
                        "items": {
                            "type": "string"
                        },
                        "title": "Funding Percentage",
                        "type": "array"
                    },
                    "roles_root": {
                        "default": [],
                        "items": {
                            "type": "string"
                        },
                        "title": "Roles Root",
                        "type": "array"
                    },
                    "roles": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RoleFormErrorsSchema"
                        },
                        "default": {},
                        "title": "Roles",
                        "type": "object"
                    },
                    "shares_root": {
                        "default": [],
                        "items": {
                            "type": "string"
                        },
                        "title": "Shares Root",
                        "type": "array"
                    },
                    "shares": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ShareFormErrorsSchema"
                        },
                        "default": {},
                        "title": "Shares",
                        "type": "object"
                    },
                    "items": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/EntryItemErrorsSchema"
                        },
                        "default": {},
                        "title": "Items",
                        "type": "object"
                    }
                },
                "title": "EntryFormErrorsSchema",
                "type": "object"
            },
...
            "EntryFormSchema": {
                "properties": {
                    "estimated_total": {
                        "title": "Estimated Total",
                        "type": "integer"
                    },
                    "funding_project_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Funding Project Id"
                    },
                    "funding_percentage": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Funding Percentage"
                    },
                    "roles": {
                        "items": {
                            "$ref": "#/components/schemas/RoleFormSchema"
                        },
                        "title": "Roles",
                        "type": "array"
                    },
                    "shares": {
                        "items": {
                            "$ref": "#/components/schemas/ShareFormSchema"
                        },
                        "title": "Shares",
                        "type": "array"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/EntryItemSchema"
                        },
                        "title": "Items",
                        "type": "array"
                    }
                },
                "required": [
                    "estimated_total",
                    "funding_project_id",
                    "funding_percentage",
                    "roles",
                    "shares",
                    "items"
                ],
                "title": "EntryFormSchema",
                "type": "object"
            },
...
            "EntryRoleSchema": {
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "value": {
                        "title": "Value",
                        "type": "integer"
                    },
                    "role_approximate_percentage": {
                        "title": "Role Approximate Percentage",
                        "type": "number"
                    }
                },
                "required": [
                    "name",
                    "value",
                    "role_approximate_percentage"
                ],
                "title": "EntryRoleSchema",
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "CanAccessPVE": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "NeedsPermission": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    },
    "servers": []
}